Routes are in a seperate file to clean up app.js
